be fully event-driven instead of poll-driven
The method that uses idle handlers really is not compatible with
the event-driven model (push a frame when it arrives). It leads
to 100% CPU as the gstreamer loop is desesperately trying to poll
the input port.

Instead, move the gstreamer loop in a separate thread and push data
when it arrives on the task.