fix issues related to the GLib main loop and threads
GLib has a concept of a thread-local main loop, which is automatically
assigned and can be automatically used (in our case, by libsoup)

Because Qt is integrated with GLib, the Qt main loop may allocate
and assign this main loop, breaking our code. Start the whole glib/libsoup
in a thread we control to avoid the problem.