fix: force reconnection if the image parameters change
WebRTC does not allow for dynamic renegotiation. The general mechanism
seems to be to create a parallel stream and let the client pick it before
we shut down the first one.

However, this mechanism would not help in our case since what changed
is the source. We could rescale to transition cleanly, but it would
be all to handle a corner-case.